test: keep wire overlay proof within package limit
